refactor: 일급 컬렉션 적용 (#501, #519)
